% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.R
\name{AncovaDeff}
\alias{AncovaDeff}
\title{Analysis-of-Covariance (ANCOVA) Design Effect}
\usage{
AncovaDeff(r = 0, r_treat = 0)
}
\arguments{
\item{r}{The multiple correlation from the regression of the response
on covariates, not including the treatment variable, in either the two- or
one-sample design. If there is only one covariate, this reduces to the
correlation between the response and the covariate. For a two-sample
analysis, this should be estimated from the Control or baseline group data
only, so that there is no bias is estimating this parameter from the
presence of a treatment effect.}

\item{r_treat}{The multiple correlation from the regression of the
treatment variable on all the covariates, applicable only in a two-sample
analysis. This is typically close to zero and negligible for a randomized
experiment. But for an observational study, this could be large enough
so that we need to account for it.}
}
\value{
The ANCOVA design effect for the study design.
}
\description{
For a two-sample design, compute the design effect from using an analysis
of covariance (ANCOVA) relative to an analysis of variance (ANOVA) design,
using the approximation of Borm et al. (2007). The ANCOVA design is analyzed
using a regression model on treatment and additional covariates, while the
ANOVA design is analyzed using regression model on treatment assignmen
alone.
  The design effect can be used in a one-sample design as well, even though
the name ANCOVA would typically not be used in this setting. We essentially
compare the error variance of a regression model on covariates with the
error variance of a regression model with no covariates, i.e., an
intercept-only model, to obtain the design effect.
  The design effect is the multiplier for the minimum required sample size
(MRSS), and the square root of the design effect is the multiplier for
the minimum detectable effect (MDE) and the maximum margin of error (MME).
See the documentation for \code{ImbalanceDeff} for details.
  Note that for a two-sample analysis, there could be imbalances between
the Treatment and Control groups, e.g., due to the study being an
observational without randomized group assignment, or due to a low sample
size. In this setting, the design effect needs to be multiplied by the
variance inflation factor (VIF) for the treatment effect, which is a function
of the R-squared of the regression of the treatment variable on all
covariates.
}
\references{
\itemize{
Borm, George F., Jaap Fransen, and Wim A. J. G. Lemmens (2007). "A Simple
  Sample Size Formula for Analysis of Covariance in Randomized Clinical
  Trials." \emph{Journal of Clinical Epidemiology,} 60: 1234-1238.
}
}
