% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_prediction.R
\name{ExtendData}
\alias{ExtendData}
\title{Extend Data}
\usage{
ExtendData(
  data,
  extend_vars = "days",
  group_vars = NULL,
  other_vars = NULL,
  horizon = 1:60
)
}
\arguments{
\item{data}{A data frame containing the variables to extend and, optionally,
the grouping variables and other variables.}

\item{extend_vars}{A character vector containing names of variables to extend
over the specified time horizon, typically, the time variables used in the
forecast exercise, e.g., the number of days, date, etc.}

\item{group_vars}{A character vector containing names of the grouping
variables. These can be considered the keys for the data that we group by.}

\item{other_vars}{A character vector containing names of other measure
variables to include in the data frame returned. These will have \code{NA}
values for time points in the extension.}

\item{horizon}{A numeric vector of positive integers to add to the time
variables to extend. Defaults to \code{1:60}.}
}
\value{
The original data frame passed into \code{data} but with additional
  rows added for the forecast horizon.
}
\description{
Extend a data frame containing time variables, grouping variables, and
other variables over a longer time horizon, so that it can be used for
forecasting, e.g., forecasting the growth of sample size counts.
}
