% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{ICC}
\alias{ICC}
\title{Intra-cluster Correlation}
\usage{
ICC(data, y, cluster_id = NULL)
}
\arguments{
\item{data}{A data frame containing the response variable in \code{y} and
and the cluster id variable in \code{cluster_id}.}

\item{y}{The string name of the response variable.}

\item{cluster_id}{The string name of the cluster id variable.}
}
\value{
The ICC of the response variable \code{y} given the clustering
  specified by \code{cluster_id}.
}
\description{
Compute the intra-cluster correlation (ICC) for a simple random intercept
model without covariates.
}
