% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.R
\name{ImbalanceDeff}
\alias{ImbalanceDeff}
\title{Imbalance Design Effect}
\usage{
ImbalanceDeff(phi = 0.5)
}
\arguments{
\item{phi}{The proportion of observations in one of the arms, either
Treatment or Control, for a two-sample analysis. Defaults to \code{0.50},
i.e., a balanced allocation for both arms.}
}
\value{
The imbalance design effect for the study with two arms.
}
\description{
Compute the design effect (relative variance) from an imbalance in number of
observations between two study arms or groups, typically, Treatment and
Control, relative to a balanced design.
  Given \code{phi} as the proportion in one of the two arms, the design
effect is given by \code{1 / (4 * phi * (1 - phi))}, which is applied as a
post-hoc multiplier in two ways:
(1) The design effect is multiplied to the minimum required sample size
    (MRSS), which itself is a relative variance measure.
(2) The square root of the design effect, a measure of relative standard
    standard deviation, is multiplied to the minimum detectable effect (MDE)
    and the maximum margin of error (MME), which themselves are relative
    standard deviation measures.
}
