% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{LogLikFun}
\alias{LogLikFun}
\title{Log-likelihood Function}
\usage{
LogLikFun(dist = c("norm", "logis", "cauchy", "lnorm", "gamma"))
}
\arguments{
\item{dist}{One of the distribution keyword strings specified above.}
}
\value{
A function taking a vector \code{par} of parameters and a vector
  \code{y} of observed data to compute the log-likelihood of the observed
  data in \code{y} for the parametric distribution specified in \code{dist}.
}
\description{
Given a variable \code{y} in the parent environment (not passed into this
function), return an R function that computes the log-likelihood of \code{y}
for one of the following parametric distributions, specified in the argument
\code{dist}:
(1) One of three symmetric distributions: the normal (\code{"norm"}),
    logistic (\code{"logis"}), and Cauchy (\code{"cauchy"}) distributions, or
(2) One of two right-skewed distributions: the lognormal (\code{"lnorm"}) and
    gamma (\code{"gamma"}) distributions.
}
