% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.R
\name{MdePropTest}
\alias{MdePropTest}
\title{Minimum Detectable Effect for the Binomial Proportion Test}
\usage{
MdePropTest(
  n = NULL,
  p0 = NULL,
  sig_level = 0.05,
  power = 0.8,
  deff = 1,
  alternative = c("two_sided", "one_sided"),
  num_samples = c("two_sample", "one_sample"),
  effect_type = c("relative", "absolute"),
  method = c("unconditional", "conditional", "asin"),
  estimation = c("analytic", "numerical"),
  analysis = c("power", "precision")
)
}
\arguments{
\item{n}{The number of observations (per group).}

\item{p0}{The binomial proportion in the baseline group for a two-sample
analysis, or a binomial proportion under the null hypothesis for a
one-sample analysis.}

\item{sig_level}{The significance level (Type I error probability).}

\item{power}{The power of the test (1 minus Type II error probability).}

\item{deff}{The design effect applied to the sample size. Defaults to 1.}

\item{alternative}{Either one- or two-sided test. Can be abbreviated.}

\item{num_samples}{Either one- or two-sample analysis.}

\item{effect_type}{Whether a relative effect or an absolute effect.}

\item{method}{The method for inference, either the unconditional,
conditional, or arcsine transformation method.}

\item{estimation}{The estimation procedure, either using an analytic formula
or a numerical root-finding procedure. The conditional method currently
implements the numerical procedure only, without the option for an
analytic formula.}

\item{analysis}{The analysis type, whether a power analysis or a precision
analysis. Note that the precision analysis is currently available only for
the unconditional and arcsine methods using the analytic estimation
procedure.}
}
\value{
The MDE, either the absolute MDE or the minimum detectable relative
  effect (MDRE).
}
\description{
Computes the minimum detectable effect (MDE) for a binomial proportion test
with the following options:
(1) A relative effect or an absolute effect.
(2) The unconditional, conditional, and arcsine transformation methods.
(3) Estimation using analytic formulas or numerical root-finding procedures.
(4) Power analysis or precision analysis.
(5) Two- or one-sample analysis.
(6) Two- or one-sided test.
}
