% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{Mle}
\alias{Mle}
\title{Maximum Likehood Estimates of Model Parameters}
\usage{
Mle(
  y,
  dist = c("norm", "lnorm", "logis", "cauchy", "gamma"),
  sd_mle = TRUE,
  ...
)
}
\arguments{
\item{y}{A numeric vector containing the observed data for estimating the
MLE of the parameters.}

\item{dist}{One of the distribution keyword strings specified in the
\code{LogLikFun} function.}

\item{sd_mle}{Whether to use the MLE for the sample standard deviation and
variance estimators in the start values. See documentation in the
\code{StartPars} function for details.}

\item{...}{Additional arguments to pass into the \code{base::optim} function,
e.g., the \code{method} argument for the optimization method. See
documentationn for the \code{base::optim} for details.}
}
\value{
A vector containing the MLEs of the parameters under the assumed
  parametric model.
}
\description{
Given a numeric vector \code{y}, compute the maximum likelihood estimates
(MLEs) for one of the parametric distributions specified for the
\code{LogLikFun} and \code{StartPars} functions. See documentation for these
two functions for details.
}
