% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.anova.oneway.R
\name{power.anova.oneway}
\alias{power.anova.oneway}
\title{Power calculations for balanced one-way analysis of variance tests}
\usage{
power.anova.oneway(groups = NULL, n = NULL, f = NULL, power = NULL,
  sig.level = 0.05)
}
\arguments{
\item{groups}{Number of groups}

\item{n}{Number of observations (per group)}

\item{f}{Effect size, Cohen's f}

\item{power}{power of study}

\item{sig.level}{significance level}
}
\description{
Power calculations for balanced one-way analysis of variance tests
}
\examples{
power.anova.oneway(groups=4,n=20,f=0.28)
}
\seealso{
\code{\link{ES.anova.oneway}}
}

