% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{PowerAnalysis}
\alias{PowerAnalysis}
\alias{PowerAnalysis-package}
\title{PowerAnalysis: Power and Precision Analysis to Determine Minimum Sample Size and Effect}
\description{
Perform power analysis and/or precision analysis for sample means of continuous and binary metrics and sample quantiles of continuous metrics, assuming asymptotic normality. One can determine the minimum required sample size (MRSS), the minimum required duration (MRD), or the minimum detectable effect (MDE) and account for design effects due to arm imbalance, analysis of covariance (ANCOVA), and cluster randomization. There are also helper functions to forecast sample size growth over time to convert between sample size growth and duration.
}
\author{
\strong{Maintainer}: Shyue-Ming Loh \email{shyuemingloh@gmail.com}

Authors:
\itemize{
  \item Randy Lai \email{randylai@google.com}
  \item Meeyoung Park \email{meeyoung@google.com}
}

}
