% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerANOVA.R
\name{powerANOVA}
\alias{powerANOVA}
\title{Returns the statistical power for a multi-factorial ANOVA}
\usage{
powerANOVA(studyDesign, N, means, SD, labels, wcorr, alpha)
}
\arguments{
\item{studyDesign}{an string including the study design, e.g. 2b*4w}

\item{N}{sample size in each condition}

\item{means}{Vector specifying mean for each condition}

\item{SD}{estimated pooled standard deviation}

\item{labels}{string array list to specify the factor names and conditions}

\item{wcorr}{correlation between dependent variables (single value or matrix)}

\item{alpha}{alpha criterion for rejecting the null hypothesis}
}
\value{
The power analysis results for ANOVA level effects
}
\description{
This function takes the Superpower package to return N of the estimates size
}
