% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.R
\name{PowerPrecisionAnalysis}
\alias{PowerPrecisionAnalysis}
\title{Power and Precision Analysis}
\usage{
PowerPrecisionAnalysis(
  stat0,
  sd0 = NULL,
  cor0 = 0,
  cor0_treat = 0,
  n = 10000,
  effect = 0,
  estimand = c("mrss", "mde"),
  analysis = c("power", "precision"),
  resp_type = c("binary", "continuous"),
  stat_type = c("mean", "quantile"),
  effect_type = c("relative", "absolute", "effect_size"),
  method = c("unconditional", "conditional", "asin"),
  alternative = c("two_sided", "one_sided"),
  num_samples = c("two_sample", "one_sample"),
  sig_level = 0.05,
  power = 0.8,
  phi = 0.5,
  cluster_stats = list(mean = 1, sd = 0, icc = 0),
  cluster_deff_type = c("min", "max")
)
}
\arguments{
\item{stat0}{A numeric vector of the sample statistics for various
conversion metrics, either sample means when \code{stat_type == "mean"} or
sample quantiles when \code{stat_type == "quantile"}. For a one-sample
analysis, the vector entries are the assumed values of the underlying
parameter, either mean or quantile, under the null hypothesis, instead of
the sample statistic. For a two-sample analysis, these are values of the
sample statistic as estimates of the underlying parameter of the Control
group population.}

\item{sd0}{A numeric vector of the estimates of population standard
deviations (SDs) for various conversion metrics, used only when
\code{resp_type == "continuous"}, where we assume that the population SD is
independent from the mean when \code{stat_type == "mean"} and the
population SD is independent from the population quantile when
\code{stat_type == "quantile"}. This could hold for quantiles, because
we assume that changes to the quantile are due location shifts to the
distribution, so that density or sparsity evaluated at the quantile of
interest stays the same. When \code{resp_type == "binary"}, we compute
the population SD in the function instead, because it is dependent on the
value of \code{effect}.}

\item{cor0}{A numeric vector of multiple correlation R values for the
regression of the corresponding conversion on covariates. These could be
the correlations between the post-period and pre-period metrics for the
conversion for the regression of the post-period conversion on the
pre-period conversion. But they could be the square-root of the R-squared
values for the regression of the conversion on any set of covariates.
See documentation for \code{AncovaDeff} for details.}

\item{cor0_treat}{A numeric vector of multiple correlation R values for the
regression of the treatment variable on covariates, applicable only in a
two-sample analysis, typically when there are imbalances between the
study groups. See documentation for \code{AncovaDeff} for details.}

\item{n}{A numeric vector of MRSS values to determine the corresponding MDEs
in a power analysis or precision analysis.}

\item{effect}{A numeric vector of MDE values to determine the corresponding
MRSS values in a power analysis, or a numeric vector of maximum
margin-of-error (MME) values in a precision analysis.}

\item{analysis}{The type of analysis conducted, either \code{"power"} for
power analysis or \code{"precision"} for precision analysis.}

\item{resp_type}{The type of response variable for the conversion metrics,
either \code{"binary"} for binary metrics or \code{"continuous"} for
continuous metrics.}

\item{effect_type}{The type of effect for the MDE, either \code{"absolute"}
for an absolute effect, \code{"relative"} for a relative effect, and
\code{"effect_size"} for an effect size.}

\item{method}{The method used for the power analysis and/or precision
analysis, either \code{"unconditional"} for the unconditional method,
\code{"conditional"} for the conditional method, and \code{"asin"} for
arcsine transformation method. Note that the unconditional method is
available for all use cases. The conditional method is valid only for
performing power analysis on conversions with binary responses. We
allow for determining the MRSS only with the conditional method. While we
we can determine the MDE as well using the conditional method, this feature
is currently not implemented yet but may be implemented in the future.
The arcsine method is valid is valid only for conversions with binary
responses. We currently don't offer this method for precision analysis
or for determining the MDE. This may change in the future.}

\item{alternative}{Whether the underlying hypothesis test for the power
precision analysis is one- or two-sided. Defaults to \code{"two_sided"} for
a two-sided test.}

\item{num_samples}{Either \code{"two_sample"} for a two-sample analysis or
\code{"one_sample"} for a one-sample analysis. Note that the two-sample
analysis is more common than the one-sample analysis, but we include the
latter for completeness' sake.}

\item{sig_level}{The significance level for the power or precision analysis.
Defaults to \code{0.05}.}

\item{power}{The minimum power required for the power analysis. Defaults to
\code{0.80}.}

\item{phi}{The proportion of observations in one of the arms, either
Treatment or Control, for a two-sample analysis. Defaults to \code{0.50},
i.e., a balanced allocation for both arms. This determines the design
effect for arm imbalance, given by \code{1 / (4 * phi * (1 - phi))}. For
details, see documentation for the \code{ImbalanceDeff} function.}

\item{cluster_stats}{A named list of statistics for evaluating the cluster
design effect, with \code{mean} denoting the mean of cluster sizes,
\code{sd} denoting the standard deviation (SD) of cluster sizes, and
\code{icc}, the intra-cluster correlation (ICC), i.e., the within-cluster
correlation. See documentation for \code{ClusterDeff} for details.}

\item{cluster_deff_type}{The type of cluster design effect, either
\code{"min"} for the lower bound or \code{"max"} for the upper bound.
See documentation for the \code{ClusterDeff} function for details.}

\item{stat}{The power analysis or precision analysis statistic to determine,
either \code{"mrss"} for the MRSS or \code{"mde"} for the MDE.}
}
\value{
Either a numeric vector of MRSS values or MDE/MME values.
}
\description{
Conduct a power analysis or precision analysis to determine the minimum
required sample size (MRSS) given a minimum detectable effect (MDE), or
vice versa, i.e., determine the MDE given an MRSS. Note that in the context
of a precision analysis, the MDE should be interpreted as the maximum error
margin (MME) instead of a detectable effect.
  While the name of the function \code{PowerPrecisionAnalysis} seems
long-winded, we typically use one of the many subvariants of the function,
as follows:
(1) \code{PowerAnalysis} and \code{PrecisionAnalysis} for, respectively,
    power analysis and precision analysis.
(2) \code{MinRequiredSampleSize}, \code{MinSampleSize}, or \code{MRSS} for
    performing a power/precision analysis to determine the MRSS.
(3) \code{MinDetectableEffect} or \code{MDE} for performing a power/precision
    analysis to determine the MDE.
  Note the following convention for the MRSS values:
(1) For the one-sample case, the MRSS values represent the sample size, i.e.,
    the total number of observations from the single sample.
(2) For the two-sample case, the MRSS values are in terms of the total number
    of observations (from the two samples) divided by two. For a balanced
    allocation, this is simply the number of observations per arm, for the
    two arms. For an unbalanced allocation, we can think of this as the
    average number of observations per arm.
}
\references{
\itemize{
\item Chow, Shien-Chung, Jun Shao, and Hansheng Wang (2003). \emph{Sample
  Size Calculations in Clinical Research.} Marcel Dekker: New York.
}
}
