% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.R
\name{PowerPrecisionAnalysisData}
\alias{PowerPrecisionAnalysisData}
\alias{PowerAnalysisData}
\alias{PrecisionAnalysisData}
\title{Power and Precision Analysis Data}
\usage{
PowerPrecisionAnalysisData(
  data,
  y_binary = NULL,
  y_continuous_mean = NULL,
  y_continuous_quantile = NULL,
  ancova_deff = FALSE,
  x = NULL,
  include_pre = FALSE,
  post_suffix = "_post",
  pre_suffix = "_pre",
  cluster_deff = FALSE,
  cluster_id = NULL,
  quantile_probs = c(0.5, 0.9, 0.95, 0.99),
  quantile_type = 7,
  all_exposure = "all",
  ...
)

PowerAnalysisData(...)

PrecisionAnalysisData(...)
}
\arguments{
\item{data}{A data frame containing disaggregated data to estimate inputs
for power or precision analysis.}

\item{y_binary}{A named list of numeric vectors containing the names of the
binary metrics.}

\item{y_continuous_mean}{A named list of numeric vectors containing the names
of continuous metrics that use a mean statistic.}

\item{y_continuous_quantile}{A named list of numeric vector containing the
names of continuous metrics that use a quantile statistic.}

\item{ancova_deff}{Whether the statistics to estimate the analysis of
covariance (ANCOVA) design effect, namely, the regression-adjusted standard
deviation \code{sd_adj} and the regression multiple correlation \code{r},
should be computed.}

\item{x}{A numeric vector containing the names of all covariates used
for the regression adjustment of all metrics.}

\item{include_pre}{Whether to include the pre-period version of the metrics
as covariates in addition to those specified in \code{x}.}

\item{post_suffix}{The suffix used for the metrics to indicate that they
measured in the post-period. These would be replaced by \code{pre_suffix}
to obtain the names of the pre-period metrics if
\code{include_pre = TRUE}.}

\item{pre_suffix}{The suffix used for the names of the pre-period versions
of the metrics.}

\item{cluster_deff}{Whether the statistics to estimate the cluster design
effect, namely the cluster size mean \code{cs_mean}, the cluster size
standard deviation \code{cs_sd}, and the intracluster correlation
\code{icc}, should be computed.}

\item{cluster_id}{String name of the cluster ID variable.}

\item{quantile_probs}{The quantile probabilities for the quantiles to
conduct power or precision analysis for.}

\item{quantile_type}{The type of quantile estimated for the \code{stat_value}
column of the data frame, basically the value passed into the \code{type}
argument of the \code{stats::quantile} function.}

\item{all_exposure}{String keyword to refer to using all observations as
the exposure condition.}

\item{...}{Additional arguments passed into the \code{QuantileSD} function
to estimate the quantile population SD.}
}
\value{
A data frame containing the key columns mentioned in the description
  above.
}
\description{
Preprocess data for power or precision analysis, with the following key
variables:
(1) Name of conversion metric.
(2) Response type, whether binary or continuous.
(3) Statistic type, whether mean or a quantile.
(4) Number of observations for the power or precision analysis.
(5) Value of the statistic for the analysis, either mean or quantile.
(6) Population standard deviation, without covariate adjustment.
(7) Population standard deviation, with covariate adjustment.
(8) Multiple correlation for the regression of the metric on the covariates,
    available only for an analysis using the mean statistic.
(9) Optionally, statistics for estimating the design effect for clustered
    data, including the mean and standard deviation of the cluster sizes
    and the estimated intra-cluster correlation (ICC) for the metric given
    the clusters.
  The arguments \code{y_binary}, \code{y_continuous_mean}, and
\code{y_continuous_quantile} take named list of numeric vector containing
the names of, respectively, binary metric, continuous metrics using the
mean statistic, and continuous metrics using the quantile statistic.
The list names are names of exposure variables. The exposure name specified
in the argument \code{all_exposure} is a keyword for using all observations.
  Note that we have three different names for this function, i.e.,
\code{PowerPrecisionAnalysisData}, \code{PowerAnalysisData}, and
\code{PrecisionAnalysisData}. All three are identical, since the input data
prepared for power analysis is the same as that for precision analysis.
}
