% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.R
\name{PowerPrecisionAnalysisTable}
\alias{PowerPrecisionAnalysisTable}
\alias{DurationTable}
\alias{MdeTable}
\alias{MaxErrorMarginTable}
\title{Power and Precision Analysis Table}
\usage{
PowerPrecisionAnalysisTable(
  data,
  exposure = "exposure",
  metric = "metric",
  resp_type = "resp_type",
  stat_type = "stat",
  n0 = "n",
  stat0 = "stat_value",
  sd0 = "sd",
  cor0 = "r",
  table_type = c("duration", "mde", "mme"),
  analysis = c("power", "precision"),
  effect = seq(0.02, 0.08, by = 0.02),
  duration = seq(4, 16, by = 4),
  suffix = NULL,
  round_up = TRUE,
  ancova_deff = FALSE,
  cluster_deff = FALSE,
  cluster_deff_type = c("min", "max"),
  cs_mean = "cs_mean",
  cs_sd = "cs_sd",
  icc = "icc",
  arms = 2,
  ...
)

DurationTable(...)

MdeTable(...)

MaxErrorMarginTable(...)
}
\arguments{
\item{data}{A data frame containing inputs for the power analysis, i.e.,
columns for the metric names and the metric means, standard deviations, and
ANCOVA/pre-post correlations.}

\item{exposure}{A string containing the name of the column of exposure
variables.}

\item{metric}{A string containing the name of the column of names for
various conversion metrics.}

\item{resp_type}{A string containing the name of the column of defining the
response type, whether binary or continuous. See documentation for
\code{PowerPrecisionAnalysis} for details.}

\item{stat_type}{A string containing the name of the column of defining the
statistic type, whether a mean or a quantile. See documentation for
\code{PowerPrecisionAnalysis} for details.}

\item{n0}{A string containing the name of the column for the number of
exposures (traffic size) for fixed time period, e.g., 4 weeks.}

\item{stat0}{A string containing the name of the column of statistics for the
conversion metrics.}

\item{sd0}{A string containing the name of the column of standard deviations
for the conversion metrics.}

\item{cor0}{A string containing the name of the column of multiple
correlation R values for the regression of the corresponding conversion on
covariates. These could be the correlations between the post-period and
pre-period metrics for the conversion for the regression of the post-period
conversion on the pre-period conversion. But they could be the square-root
of the R-squared values for the regression of the conversion on any set of
covariates.}

\item{table_type}{A string specifying the table type, whether
\code{"duration"} for a table of minimum durations or \code{"mde"} for a
table of MDEs or MMEs.}

\item{analysis}{The type of analysis conducted, either \code{"power"} for
power analysis or \code{"precision"} for precision analysis. This currently
applies only to the duration table, i.e., \code{table_type = "duration"},
since \code{table_type = "mde"} and \code{table_type = "mme"} will
specify the type of analysis by their argument values alone, i.e., a power
analysis for \code{table_type = "mde"} and a precision analysis for
\code{table_type = "mme"}.}

\item{effect}{A numeric vector of MDE values to determine the corresponding
MRSS values in a power analysis or MME values for a precision analysis.}

\item{duration}{A numeric vector of durations (in weeks) to determine the
corresponding MDE values in a power analysis or a precision analysis.}

\item{suffix}{A numeric vector of suffix values for the duration/MDE column
names. Defaults to the duration/MDE sequence of values if \code{NULL}.}

\item{round_up}{Whether to round up the durations in a table of durations.}

\item{ancova_deff}{Whether the analysis of covariance (ANCOVA) design effect
should be applied.}

\item{cluster_deff}{Whether the cluster design effect should be applied.}

\item{cluster_deff_type}{The type of cluster design effect to apply, whether
\code{"min"} for the lower bound or \code{"max"} for the upper bound.\
See documentation for the \code{ClusterDeff} function for details.}

\item{cs_mean}{Name of column with cluster size means.}

\item{cs_sd}{Name of column with cluster size standard deviations.}

\item{icc}{Name of column containing estimated intra-cluster correlations
(ICCs).}

\item{arms}{The number of arms in the study.}

\item{...}{Additional arguments to pass into the \code{MRSS} or \code{MDE}
functions.}
}
\value{
A data frame containing the duration or MDE table, with percent
  format in relevant columns.
}
\description{
Create table of power/precision analysis results, either a table of
minimum durations, or a table of minimum detectable effects (MDEs) for a
power analysis or a table of maximum margins of error (MMEs) for a precision
analysis.
}
