% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.R
\name{PowerPropTest}
\alias{PowerPropTest}
\title{Power Analysis for the Binomial Proportion Test}
\usage{
PowerPropTest(
  n = NULL,
  p1 = NULL,
  p2 = NULL,
  sig_level = 0.05,
  power = NULL,
  deff = 1,
  alternative = c("two_sided", "one_sided"),
  num_samples = c("two_sample", "one_sample"),
  strict = FALSE,
  method = c("conditional", "unconditional"),
  tol = .Machine$double.eps^0.25
)
}
\arguments{
\item{n}{The number of observations (per group).}

\item{p1}{The probability in one group for a two-sample analysis, or the
probability under the null hypothesis for a one-sample analysis.}

\item{p2}{The probability in other group for a two-sample analysis, or the
probability under the alternative hypothesis for a one-sample analysis.}

\item{sig_level}{The significance level (Type I error probability).}

\item{power}{The power of the test (1 minus Type II error probability).}

\item{deff}{The design effect applied to the sample size. Defaults to 1.}

\item{alternative}{Either one- or two-sided test. Can be abbreviated.}

\item{num_samples}{Either one- or two-sample analysis.}

\item{strict}{Use strict interpretation in two-sided case.}

\item{method}{The method for inference, either conditional or unconditional.}

\item{tol}{The numerical tolerance used in root finding, the default
providing (at least) four significant digits.}
}
\value{
Object of class "power.htest", a list of the arguments (including the
  computed one) augmented with method and note elements.
}
\description{
An enhanced version of the \code{power.prop.test} function in base R, which
according to its documentation functions to: "Compute the power of the
two-sample test for proportions, or determine parameters to obtain a target
power." The \code{power.prop.test} function uses the conditional method,
but \code{PowerPropTest} function enhances it to allow for the unconditional
method, as well as a one-sample test. The code for the most part is taken
from \code{power.prop.test} but with some modifications for syntax and
implementing the unconditional method. Similarly, the following documentation
has been taken almost verbatim from that of \code{power.prop.test}.
}
