% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerRegression.R
\name{powerRegression}
\alias{powerRegression}
\title{Returns the sample size for a one-way ANOVA}
\usage{
powerRegression(ivs, N, fsquared, alpha)
}
\arguments{
\item{ivs}{number of independent variables}

\item{N}{number of samples}

\item{fsquared}{estimated effect size as f-squared}

\item{alpha}{alpha criterion for rejecting the null hypothesis}
}
\value{
The sample size
}
\description{
This function takes the power package to return only N of the estimates size
}
