% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{QuantileCI}
\alias{QuantileCI}
\title{Confidence Interval for a Sample Quantile}
\usage{
QuantileCI(
  y,
  tau = 0.5,
  method = c("interpolate", "exact", "normal_approx"),
  alternative = c("two_sided", "one_sided"),
  conf_level = 0.95,
  min_coverage = TRUE
)
}
\arguments{
\item{y}{A numeric vector containing observed data to estimate the sample
quantile CI for.}

\item{tau}{The quantile probability for the sample quantile.}

\item{method}{One of the three methods for obtaining the nonparametric
quantile CI, i.e., \code{"interpolate"} for the exact method with
interpolation, \code{"exact"} for the exact method without interpolation,
and \code{"normal_approx"} for using an asymptotic normal approximation for
the binomial CI.}

\item{alternative}{Whether the CI corresponds to one- or two-sided hypothesis
test. Defaults to \code{"two_sided"} for a two-sided test.}

\item{conf_level}{The confidence level for the CI. This is approximate for
the exact method.}
}
\value{
A two-entry numeric vector containing lower and upper bounds for the
  sample quantile CI estimate.
}
\description{
Estimate the confidence interval (CI) of a sample quantile using order
statistics, with the following three variants:
(1) Asymptotic normal approximation using uninterpolated order statistics.
(2) Exact method using uninterpolated order statistics.
(3) Exact method using interpolated order statistics, first proposed by
    Hettmansperger and Sheather (1986) for the median and later extended to
    arbitrary quantiles by Nyblom (1992).
  This is currently implemented as a wrapper function around the following
three functions in the \code{EnvStats} that computes the sample quantile CIs
for, respectivel, the above three methods:
(1) \code{EnvStats::ci.qnpar.normal.approx} for the normal approximation
    method.
(2) \code{EnvStats::ci.qnpar.exact} for the exact method without
    interpolation.
(3) \code{EnvStats::ci.qnpar.interpolate} for the exact method with
    interpolation.
}
\references{
\itemize{
\item Hettmansperger, Thomas P., and Sheather, Simon J. (1986). "Confidence
  Intervals Based on Interpolated Order Statistics." \emph{Statistics &
  Probability Letters,} 4:75-79.
\item Nyblom, Jukka (1992). "Note on Interpolated Order Statistics."
  \emph{Statistics & Probability Letters,} 14: 129-131.
}
}
