% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{QuantileSD}
\alias{QuantileSD}
\title{Standard Deviation and Standard Error for a Sample Quantile}
\usage{
QuantileSD(y, tau = 0.5, ...)
}
\arguments{
\item{y}{A numeric vector containing the observed response values or
quantile regression residuals for estimating the asymptotic SD or SE.}

\item{...}{Additional arguments to pass into the \code{Sparsity} function for
the \code{QuantileSD} function, or into the \code{QuantileSD} function for
the \code{QuantileSE} function.}
}
\value{
A scalar estimate of the asymptotic population SD of the sample
  quantile for the \code{QuantileSD} function and the asymptotic SE for the
  \code{QuantileSE} function.
}
\description{
Given a numeric vector \code{y} of observations and \code{tau} for the
quantile probability, estimate the following asymptotic measures:
(1) The asymptotic population standard deviation (SD) for the sample
    \code{tau}-th quantile, given by
    \code{tau * (1 - tau) * (1 / f(Q(tau)))}, where \code{1 / f(Q(tau))} is
    the sparsity evaluated at the quantile of interest.
(2) The asymptotic standard error (SE) for the sample quantile, defined as
    the asymptotic population SD divided by \code{sqrt(n)}, i.e., the
    square root of the sample size.
  Note that the asymptotic population SD of the sample quantile is denoted by
\code{omega} in Equation 3.2.1 in Koenker (2005). See documentation for the
\code{Sparsity} function for details.
}
\references{
\itemize{
\item Koenker, Roger (2005). \emph{Quantile Regression}. Cambridge University
  Press.
}
}
