% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSize.R
\name{sampleSize}
\alias{sampleSize}
\title{Returns the sample size for a one-way ANOVA}
\usage{
sampleSize(grps, SD, pwr, alpha)
}
\arguments{
\item{grps}{an array of conditions}

\item{SD}{estimated pooled standard deviation}

\item{pwr}{estimated statistical power}

\item{alpha}{alpha criterion for rejecting the null hypothesis}
}
\value{
The sample size
}
\description{
This function takes the power package to return only N of the estimates size
}
