% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_prediction.R
\name{SampleSizeData}
\alias{SampleSizeData}
\title{Sample Size Data}
\usage{
SampleSizeData(
  data,
  unit = "obscured_gaia_id",
  group = "experiment_id",
  x = "min_exposure_date",
  compute_days = TRUE,
  ref_date = NULL
)
}
\arguments{
\item{data}{A data frame containing the unit variable (e.g., user id),
a grouping variable, and a date variable.}

\item{unit}{String name of unit table.}

\item{group}{String name of grouping variable.}

\item{x}{String name of the variable for the x-axis of the plot, either
a character or Date vector for dates, or a numeric vector, e.g., for
number of days.}

\item{compute_days}{Whether to compute the number of days from a reference
date if the \code{x} argument is a date or character vector.}

\item{ref_date}{The reference date for computing the number of days. If
\code{NULL}, then the minimum date in the \code{data} data frame is used.}
}
\value{
A data frame containing the x-axis variable(s), grouping variables,
  and the sample size counts and cumulative counts for each combination of
  x variable(s) and grouping variables.
}
\description{
Create a data frame of sample counts and cumultative counts over time, which
be used for aggregating raw data for sample size plots using
\code{SampleSizePlot}.
}
