% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.R
\name{SampleSizeDurationConversion}
\alias{SampleSizeDurationConversion}
\alias{SampleSizeToDuration}
\alias{DurationToSampleSize}
\title{Sample Size and Duration Conversion}
\usage{
SampleSizeDurationConversion(
  n = NULL,
  duration = NULL,
  n0 = 10000,
  t0 = 4,
  arms = 2,
  traffic_prop = StudyTrafficProportion(),
  round_up = FALSE
)

SampleSizeToDuration(n, ...)

DurationToSampleSize(duration, ...)
}
\arguments{
\item{n}{A numeric vector of sample sizes. If not \code{NULL}, then the
sample size values will be converted to duration values.}

\item{duration}{A numeric vector of duration in weeks. If \code{n} is
\code{NULL} and \code{duration} is not \code{NULL}, the duration values
will be converted to sample size values.}

\item{n0}{The sample size at a particular nonzero time point, i.e., at
\code{t0}. In the language of Cloud experiments, this is the exposure
traffic size within a \code{t0}-week period. Defaults to 10,000 units.}

\item{t0}{The nonzero time point to observe the sample size \code{n0}.
Defaults to 4 weeks.}

\item{arms}{The number of arms for the experiment. Defaults to 2 arms.}

\item{traffic_prop}{The proportion of exposure traffic available for
experimentation. Defaults to \code{StudyTrafficProportion()}. See
documentation for \code{TrafficProportion} for details.}

\item{round_up}{Whether to round up the returned value.}
}
\value{
A numeric vector of sample size or duration values, depending on
  on the values of \code{n} and \code{duration}.
}
\description{
Convert from sample size to duration or vice versa, i.e., duration to
sample size, based on an assumed linear relationship between the two,
with parameters determine from interpolating and extrapolating a line
through the origin and the sample size at a nonzero time point.
}
