% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_prediction.R
\name{SampleSizeModel}
\alias{SampleSizeModel}
\title{Sample Size Model}
\usage{
SampleSizeModel(
  data,
  y,
  x,
  group = NULL,
  method = c("spline", "linear"),
  k = -1,
  bs = "tp"
)
}
\arguments{
\item{data}{A data frame containing the unit variable (e.g., user id),
a grouping variable, and a date variable.}

\item{y}{String name for the y-axis variable, which has to be a numeric
vector.}

\item{x}{String name for the x-axis variable of the plot, which has to be a
numeric vector, e.g., the number of days or a time index.}

\item{group}{String name of grouping variable. Note the function is currently
limited to using a single grouping variable. Hence, only the first entry
is used if a character vector is passed in as the argument.}

\item{method}{Method used to fit a forecast model, either a spline regression
model using \code{mgcv::gam} or a linear regression model using
\code{stats::lm}.}

\item{k}{The number of knots for the spline regression, an argument passed
into the {mgcv::s} smoothing function when \code{method = "spline"}.}

\item{bs}{The type of basis function for the spline regression, argument
into the {mgcv::s} smoothing function when \code{method = "spline"}.
Available types include \code{"tp"} for thin-plate splines and \code{"cr"}
for cubic regression splines. See the following \code{mgcv} documentation
page for other options:
https://stat.ethz.ch/R-manual/R-devel/library/mgcv/html/smooth.terms.html}
}
\value{
A \code{mgcv::gam} object or a \code{stats::lm} object for the fitted
  regression model for forecasting sample size growth.
}
\description{
Fits a forecast model for a sample size time series using \code{mgcv::gam}
for a spline regression model or \code{stats::lm} for a linear regression
model. Note the following:
(1) The spline regression model is recommended for forecasting sample size
    growth, while the linear regression model option is added as a benchmark.
(2) While the name of the function is \code{SampleSizeModel}, it can be used
    for forecasting any numeric variable \code{y} ordered by a numeric
    variable \code{x}.
}
