% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_prediction.R
\name{SampleSizePlot}
\alias{SampleSizePlot}
\title{Sample Size Plot}
\usage{
SampleSizePlot(
  data,
  unit = "obscured_gaia_id",
  y = NULL,
  y_pred = NULL,
  x = "min_exposure_date",
  facet = NULL,
  color = NULL,
  pred_range = c("horizon", "all"),
  preprocess = FALSE,
  cumulative = TRUE,
  smooth = FALSE,
  facet_ncol = 2,
  facet_scales = "free",
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{data}{A data frame containing the unit variable (e.g., user id),
a grouping variable, and a date variable.}

\item{unit}{String name of unit variable, used when the data needs to be
aggregated first by specifying \code{preprocess = TRUE}.}

\item{y}{String name of the y-axis variable to plot.}

\item{y_pred}{String name of the prediction of the y-axis variable to plot.}

\item{x}{String name of the variable for the x-axis of the plot, either
a character or Date vector for dates, or a numeric vector, e.g., for
number of days.}

\item{facet}{String name of the variable to facet the plot by.}

\item{color}{String name of the variable to color the plot by.}

\item{pred_range}{The range to show the prediction \code{y_pred}. If
\code{"horizon"}, \code{y_pred} will be shown for values of \code{x} where
the variable \code{y} has \code{NA} values. If \code{"all"}, \code{y_pred}
will be shown for all values of \code{x}.}

\item{preprocess}{Whether to aggregate the data by date and group before
passing the data to \code{ggplot2::ggplot}.}

\item{cumulative}{Whether the sample size counts are cumulative counts over
time. Defaults to \code{TRUE}.}

\item{smooth}{Whether to add a smoothed mean of the data in the plot, in
addition to the line geom of the \code{y} variable.}

\item{facet_ncol}{The number of columns for plot faceting.}

\item{xlab}{The label for the x-axis.}

\item{ylab}{The label for the y-axis.}
}
\value{
A line plot for the sample size counts, whether individual time
  counts or cumulative counts over time.
}
\description{
Create a plot of sample size counts over time using \code{ggplot2::ggplot},
optionally using \code{SampleSizeData} to create aggregated data for the
plot. We have the following dimensions for defining the plot:
(1) Data Aggregation: The function can aggregate the data using
    \code{SampleSizeData} if \code{preprocess == TRUE}. Otherwise, the
    function will take the data as is and assume that the data is already
    aggregated.
(2) Sample Size Counts: If the data needs to be aggregated first, then
    \code{y} would be set to one of two values, i.e., \code{y == "cum_count"}
    if \code{cumulative == TRUE} and \code{y == "count"} if
    if \code{cumulative == FALSE}. If the data is taken as is, then \code{y}
    should ideally be specified. If it is not, then \code{y == "cum_count"}
    or \code{y == "count"} depending on the \code{cumulative} argument.
(3) Estimated Mean: In addition to the line plot of the sample size counts,
    this function can include the line plot of estimated means of the counts,
    either provided in the data using the name provided in \code{y_pred}, or
    by adding a \code{geom_smooth()} to the plot with \code{method = "gam"}
    for a generalized additive model (GAM) smooth fit using \code{mgcv::gam}.
    The advantage of providing the estimated mean using \code{y_pred} is that
    we can plot \code{y_pred} for time points with \code{NA} values of
    \code{y}.
}
