% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_prediction.R
\name{SampleSizePrediction}
\alias{SampleSizePrediction}
\title{Sample Size Prediction}
\usage{
SampleSizePrediction(
  data,
  y = "cum_count",
  y_pred = paste0(y, "_pred"),
  x = "days",
  date = "min_exposure_date",
  group_vars = NULL,
  other_vars = "count",
  train = NULL,
  horizon = 1:60,
  ...
)
}
\arguments{
\item{data}{A data frame containing the unit variable (e.g., user id),
a grouping variable, and a date variable.}

\item{y}{String name for the y-axis variable, which has to be a numeric
vector.}

\item{y_pred}{String name for the prediction of the y-axis variable. Defaults
to \code{y} followed by a suffix \code{"_pred"}.}

\item{x}{String name for the x-axis variable of the plot, which has to be a
numeric vector, e.g., the number of days or a time index.}

\item{date}{String name for the date variable corresponding to the x-axis
variable.}

\item{group_vars}{A character vector of names of grouping variables. If there
are multiple grouping variables, a single grouping variable will be created
as a concatenation of all the grouping variable for the
\code{SampleSizeModel} function call.}

\item{other_vars}{A character vector of names of measure variables to retain
in the prediction data. These names will be passed into the
\code{other_vars} argument of the \code{ExtendData} function call.}

\item{train}{A numeric vector for values of the \code{x} variable to train
on. Defaults to \code{NULL}, where all values of \code{x} are used. Note
in addition to this filter, all observations with \code{NA} values for
\code{y} will be removed as well.}

\item{horizon}{A numeric vector of positive integers to add to the time
variables to extend. Defaults to \code{1:60}.}

\item{...}{Additional arguments to pass into the \code{SampleSizeModel}
function.}
}
\value{
A data frame containing the variables specified in \code{y},
  \code{x}, \code{group_vars}, and \code{other_vars}, and the prediction
  variable \code{y_pred} and optionally, the combined group variable
  \code{"group"} if there are multiple grouping variables.
}
\description{
Add a prediction of sample size counts to a data frame containing sample
size data. Extend the data frame over a forecast horizon using
\code{ExtendData} as needed.
}
