% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{Sparsity}
\alias{Sparsity}
\title{Sparsity Function Evaluated at Quantile}
\usage{
Sparsity(
  y,
  tau = 0.5,
  method = c("npar_sbg", "npar_ci", "par"),
  ci_method = c("interpolate", "exact", "normal_approx"),
  dist = c("norm", "logis", "cauchy", "lnorm", "gamma"),
  p = 1,
  hs = TRUE,
  sig_level = 0.05,
  pars = NULL,
  residualize = FALSE,
  tol = .Machine$double.eps^0.5
)
}
\arguments{
\item{y}{Either numeric vector containing the observed response values or
quantile regression residuals for estimating the sparsity, or
a \code{quantreg::rq} object from which the quantile regression residuals
will be extracted. Note that for the SBG method, the numeric vector
\code{y} will need to be residualized first, i.e., converted into quantile
regression residuals. This can be done using before passing the values into
the argument \code{y} or using the \code{residualize = TRUE} option if the
\code{y} values passed into the function are not quantile regression
residuals.}

\item{tau}{The quantile probability for the estimated quantile.}

\item{method}{Either \code{"npar_sbg"} for the nonparmaetric SBG method, or
\code{"npar_ci"} for the nonparametric CI method, or \code{"par"} for the
parametric method.}

\item{ci_method}{The method for obtaining the nonparametric quantile CI,
equal to the value of the \code{method} argument of the
\code{QuantileCI} function. See the documentation for the
\code{QuantileCI} function for details.}

\item{dist}{The parametric distribution for the parametric method, equal to
the value of the \code{dist} argument for the \code{LogLikFun} function.
See the documentation for the \code{LogLikFun} function for details.}

\item{p}{The number of parameters estimated in the linear conditional
quantile of a quantile regression, if \code{y} is a vector of quantile
regression residuals.}

\item{hs}{For the SBG method (\code{method = "npar_sbg"}), whether to use the
method of Hall and Sheather (1988) to define the bandwidth to estimate
sparsity. If \code{hs = FALSE}, the Bofinger (1975) method is used.}

\item{sig_level}{The significance level used to define (1) the z quantiles
in the bandwidth selected using either the Hall and Sheather (1988) method
or the Bofinger (1975) method, for the SBG method, and (2) the confidence
level of the interval for the CI method.}

\item{residualize}{Whether to residualize \code{y} using a simple
location-model quantile regression, if \code{y} is a numeric vector.
This is needed only for the SBG method, but we provide the option for the
other methods for benchmarking purposes.}
}
\value{
A scalar estimate of the sparsity evaluated at the quantile of
  interest.
}
\description{
Given a numeric vector \code{y} of observations and \code{tau} for the
quantile probability, estimate the sparsity function, evaluated at the
\code{tau}-th quantile.
  The sparsity function, a term coined by Tukey (1965), refers the inverse of
the density function. As the density function is the slope of the
distribution function, the sparsity function, as its inverse, is the slope
of the quantile function, the inverse of the distribution function. That is,
given a density function \code{f} of distribution function \code{F}, the
sparsity function is simply \code{1 / f}, the slope of the quantile function
\code{Q}, the inverse of \code{F}.
  An important fact is that if the sample observations are iid, the
asymptotic variance of a sample quantile, a difference between two sample
quantiles (of the same probability), or any regression coefficient of a
quantile regression is proportional to the sparsity function evaluated at the
quantile of interest. In particular, it is proportional to the variance term
\code{tau * (1 - tau) * (1 / f(Q(tau)))}, where \code{1 / f(Q(tau))} is the
sparsity evaluated at the quantile of interest.
  The result for the asymptotic variance of a sample quantile for iid
has been known for some time and can be found in, e.g., Cramer (1946). The
result for any regression coefficient of a quantile regression is first shown
in Koenker and Bassett (1978). Hence, estimating the asymptotic variance of a
sample quantile, a sample quantile difference, or a coefficient estimator in
a quantile regression depends in large part on the estimation of the sparsity
\code{1 / f(Q(tau))}.
  This function offers three approaches for estimating the sparsity:
(1) A nonparametric approach using the Siddiqui-Bloch-Gastwirth (SBG) method
    of Siddiqui (1960) and Block and Gaswirth (1968), implemented for the
    \code{se = "iid"} option of the \code{quantreg::summary.rq} function.
(2) An alternative nonparametric approach based on the inferring the
    sparsity from the width of a nonparametric quantile confidence interval
    (CI), based on the CI's asymptotic properties. This idea was first
    proposed by Hao et al. (2019), and to our knowledge, there isn't any
    mention of this in existing literature. We will refer to this as the CI
    method.
(3) A parametric approach, using one of the parametric distributions defined
    in the \code{LogLikFun} function. In this approach
  We recommend using the nonparametric approach, when the underlying
distribution of the data is unknown, which is typically the case. The
parametric approach can be used if we know the underlying distribution or
as validation for the nonparametric approach. For the nonparametric
approach, we recommend using the SBG method over the CI method, especially
for small and moderate sample sizes, since the former has better properties
based on simulation exercises and is the default method used for the
\code{se = "iid"} option of the \code{quantreg::summary.rq} function for
sparsity estimation. However, when the sample size is large, e.g., >= 100K
observations, the SBG method may be computationally intensive, due to the
need to fit a quantile regression on large data. Since the difference
between the SBG and CI methods for large sample sizes may be small, one may
prefer the CI method over the SBG method due to computational speed.
}
\references{
\itemize{
\item Bloch, Daniel A., and Gastwirth, Joseph L. (1968). "On a Simple
  Estimate of the Reciprocal of the Density Function." \emph{Annals of
  Mathematical Statistics,} 39: 1083-1085.
\item Bofinger, Eve (1975). "Estimation of a Density Function Using Order
  Statistics." \emph{Australian and New Zealand Journal of Statistiscs}, 17:
  1-7.
\item Cramer, Harald (1946). \emph{Mathematical Methods of Statistics}.
  Princeton University Press.
\item Hall, Peter and Simon J. Sheather (1988). "On the Distribution of a
  Studentized Quantile." \emph{Journal of the Royal Statistical Society
  Series B}, 50: 381-391.
\item Hao, Siteng, Tianhong He, and Meeyoung Park (2019). "Statistical
  Methods of Percentile Estimation and Inference in the Page Load Latency
  Study." Google Internal Document.
\item Koenker, Roger and Gilbert Bassett, Jr. (1978). "Regression Quantiles."
  \emph{Econometrica}, 46(1): 33-50.
\item Siddiqui, M. M. (1960). "Distribution of Quantiles in Samples from a
  Bivariate Population." \emph{Journal of Research of the National Bureau of
  Standards,} 64B(3): 145-150.
\item Tukey, John Wilder (1965). "Which Part of the Sample Contains the
  Information?" \emph{Proceedings of the National Academy of Sciences of the
  United States of America}, 53(1): 127-134.
}
}
