% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.R
\name{TrafficProportion}
\alias{TrafficProportion}
\alias{StudyTrafficProportion}
\alias{LaunchTrafficProportion}
\title{Traffic Proportion for Study and Launch}
\usage{
TrafficProportion(type = c("study", "launch"))

StudyTrafficProportion()

LaunchTrafficProportion()
}
\arguments{
\item{type}{Whether a study (\code{"study"}) or launch (\code{"launch"}).}
}
\value{
The proportion of traffic available for experimentation.
}
\description{
The proportion of traffic available to a study—i.e., an A/B experiment—or a
launch. In the context of many online experiments, including those at Google,
a portion of the traffic is held back from experimentation, typically 15%, so
that the proportion available for experimentation in a study is typically
85%. For a launch, the proportion is 100%.
}
