% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{TwoSidedCI2SE}
\alias{TwoSidedCI2SE}
\title{Converting a Two-sided Confidence Interval to a Sparsity Estimate}
\usage{
TwoSidedCI2SE(ci = qnorm(0.025) * c(-1, 1), conf_level = 0.95)
}
\arguments{
\item{ci}{A two-entry numeric vector containing the limits of two-sided CI.}

\item{conf_level}{The confidence level of the two-sided CI.}

\item{se}{The value of the SE.}

\item{n}{The number of observations.}

\item{sd}{The value of the SD.}

\item{tau}{The probability for the quantile of interest.}
}
\value{
The SE for the \code{TwoSidedCI2SE} function, the SD for the
   \code{SE2SD} function, and the sparsity evaluated at the quantile of
   interest for \code{SD2Sparsity} function.
}
\description{
Three intermediate functions to convert a two-sided symmetric normal
confidence interval (CI) to a sparsity estimate:
(1) Converting from the two-sided CI to the standard error (SE)
(2) Converting from an SE to a standard deviation (SD)
(3) Converting from an asymptotic quantile SD to the sparsity evaluated at
    the quantile of interest.
}
