% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{VarStabTransf}
\alias{VarStabTransf}
\title{Variance-stabilizing Transformations}
\usage{
VarStabTransf(
  y,
  dist = c("poisson", "binomial"),
  const = NULL,
  inverse = FALSE
)
}
\arguments{
\item{y}{A numeric vector containing data to apply the VST to.}

\item{dist}{The distribution of the underlying data.}

\item{const}{The proportionality constant for the VST.}

\item{inverse}{Whether to apply the inverse of the VST instead of the VST.}
}
\value{
The VST or the inverse VST of the data in \code{x}.
}
\description{
Apply the appropriate variance-stabilizing transformation (VST) or its
inverse for specific parametric distributions, implemented for the estimation
of the following parameters:
(1) Poisson mean: The square-root transformation.
(2) Binomial proportion: The arcsine transformation, also known as the
    arcsine-square-root transformation or the angular transformation.
}
