% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.R
\name{ZQuantileMultiplier}
\alias{ZQuantileMultiplier}
\title{Z-Quantile Multiplier}
\usage{
ZQuantileMultiplier(
  analysis = c("power", "precision"),
  alternative = c("two_sided", "one_sided"),
  sig_level = 0.05,
  power = 0.8
)
}
\arguments{
\item{analysis}{The type of analysis conducted, either \code{"power"} for
power analysis or \code{"precision"} for precision analysis.}

\item{alternative}{Whether the underlying hypothesis test for the power
precision analysis is one- or two-sided. Defaults to \code{"two_sided"} for
a two-sided test.}

\item{sig_level}{The significance level for the power or precision analysis.
Defaults to \code{0.05}.}

\item{power}{The minimum power required for power analysis. Defaults to
\code{0.80}. Not applicable for precision analysis.}
}
\value{
The z-quantile multiplier for power or precision analysis.
}
\description{
Helper function to compute the multiplier of z-quantiles, i.e., quantiles of
the standard normal distribution, for computing the minimum detectable
effect (MDE) in a power analysis or the maximum margin or error (MME) in a
precision analysis, assuming asymptotic normality of the test statistic,
with options for two- and one-sided alternatives. This multiplier is squared
for computing the minimum required sample size (MRSS) in power or precision
analysis.
}
\references{
\itemize{
\item Chow, Shien-Chung, Jun Shao, and Hansheng Wang (2003). \emph{Sample
  Size Calculations in Clinical Research.} Marcel Dekker: New York.
}
}
