% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.R
\name{ZUpperSig}
\alias{ZUpperSig}
\title{Z Upper Quantile for Significance}
\usage{
ZUpperSig(alternative = c("two_sided", "one_sided"), sig_level = 0.05)
}
\arguments{
\item{alternative}{Whether the underlying hypothesis test for the power
precision analysis is one- or two-sided. Defaults to \code{"two_sided"} for
a two-sided test.}

\item{sig_level}{The significance level for the power or precision analysis.
Defaults to \code{0.05}.}
}
\value{
The z upper-quantile for the significance level.
}
\description{
Helper function to compute the standard normal upper-quantile for a
specified significance level \code{sig_level}:
(1) \code{qnorm(1 - sig_level)} for a one-sided hypothesis test, and
(2) \code{qnorm(1 - sig_level / 2)} for a two-sided hypothesis test.
}
