% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poweranalysis.R
\docType{package}
\name{poweranalysis}
\alias{poweranalysis}
\title{poweranalysis: Power Analysis for Social Science Research}
\description{
The \code{poweranalysis} package provides a collection of
interactive \code{learnr} materials for the workshop
"Power Analysis for Social Science Research". The workshop consists of
a lecture and a lab. You can call the materials using the code provided
in the examples
below.
}
\examples{
\dontrun{
  learnr::run_tutorial("01-lec", package = "poweranalysis")
  learnr::run_tutorial("01-lab", package = "poweranalysis")
}
}
\seealso{
\code{\link{learnr}}
}
